Improve SOC of raw editor
Improve separation of concerns for RawEditor by moving the code
for the text input client to a separate class, furthermore add
more comments.
